Clean up unused methods from Sources API client
Rename client to the API client and clean up methods that were move
