+++ internal/clients/sources/stubs/sources_stub.go
@@ -75,10 +75,6 @@ func WithSourcesIntegration(parent context.Context, init_store *[]sources.Source
 	return ctx
 }
 
-func (mock *SourcesIntegrationStub) GetProvisioningTypeId(ctx context.Context, reqEditors ...sources.RequestEditorFn) (string, error) {
-	return "11", nil
-}
-
 func (mock *SourcesIntegrationStub) ShowSourceWithResponse(ctx context.Context, id sources.ID, reqEditors ...sources.RequestEditorFn) (*sources.ShowSourceResponse, error) {
 	lst := *mock.store
 	return &sources.ShowSourceResponse{
